+++ library/core/src/main/java/com/google/android/exoplayer2/mediacodec/AsynchronousMediaCodecAdapter.java
@@ -49,15 +49,11 @@ import java.nio.ByteBuffer;
   public static final class Factory implements MediaCodecAdapter.Factory {
     private final Supplier<HandlerThread> callbackThreadSupplier;
     private final Supplier<HandlerThread> queueingThreadSupplier;
-    private final boolean forceQueueingSynchronizationWorkaround;
     private final boolean synchronizeCodecInteractionsWithQueueing;
 
     /** Creates a factory for codecs handling the specified {@link C.TrackType track type}. */
     public Factory(@C.TrackType int trackType) {
-      this(
-          trackType,
-          /* forceQueueingSynchronizationWorkaround= */ false,
-          /* synchronizeCodecInteractionsWithQueueing= */ false);
+      this(trackType, /* synchronizeCodecInteractionsWithQueueing= */ false);
     }
 
     /**
@@ -65,23 +61,17 @@ import java.nio.ByteBuffer;
      *
      * @param trackType One of {@link C#TRACK_TYPE_AUDIO} or {@link C#TRACK_TYPE_VIDEO}. Used for
      *     labelling the internal thread accordingly.
-     * @param forceQueueingSynchronizationWorkaround Whether the queueing synchronization workaround
-     *     will be enabled by default or only for the predefined devices.
      * @param synchronizeCodecInteractionsWithQueueing Whether the adapter should synchronize {@link
      *     MediaCodec} interactions with asynchronous buffer queueing. When {@code true}, codec
      *     interactions will wait until all input buffers pending queueing wil be submitted to the
      *     {@link MediaCodec}.
      */
-    public Factory(
-        @C.TrackType int trackType,
-        boolean forceQueueingSynchronizationWorkaround,
-        boolean synchronizeCodecInteractionsWithQueueing) {
+    public Factory(@C.TrackType int trackType, boolean synchronizeCodecInteractionsWithQueueing) {
       this(
           /* callbackThreadSupplier= */ () ->
               new HandlerThread(createCallbackThreadLabel(trackType)),
           /* queueingThreadSupplier= */ () ->
               new HandlerThread(createQueueingThreadLabel(trackType)),
-          forceQueueingSynchronizationWorkaround,
           synchronizeCodecInteractionsWithQueueing);
     }
 
@@ -89,11 +79,9 @@ import java.nio.ByteBuffer;
     /* package */ Factory(
         Supplier<HandlerThread> callbackThreadSupplier,
         Supplier<HandlerThread> queueingThreadSupplier,
-        boolean forceQueueingSynchronizationWorkaround,
         boolean synchronizeCodecInteractionsWithQueueing) {
       this.callbackThreadSupplier = callbackThreadSupplier;
       this.queueingThreadSupplier = queueingThreadSupplier;
-      this.forceQueueingSynchronizationWorkaround = forceQueueingSynchronizationWorkaround;
       this.synchronizeCodecInteractionsWithQueueing = synchronizeCodecInteractionsWithQueueing;
     }
 
@@ -111,7 +99,6 @@ import java.nio.ByteBuffer;
                 codec,
                 callbackThreadSupplier.get(),
                 queueingThreadSupplier.get(),
-                forceQueueingSynchronizationWorkaround,
                 synchronizeCodecInteractionsWithQueueing);
         TraceUtil.endSection();
         codecAdapter.initialize(
@@ -153,13 +140,10 @@ import java.nio.ByteBuffer;
       MediaCodec codec,
       HandlerThread callbackThread,
       HandlerThread enqueueingThread,
-      boolean forceQueueingSynchronizationWorkaround,
       boolean synchronizeCodecInteractionsWithQueueing) {
     this.codec = codec;
     this.asynchronousMediaCodecCallback = new AsynchronousMediaCodecCallback(callbackThread);
-    this.bufferEnqueuer =
-        new AsynchronousMediaCodecBufferEnqueuer(
-            codec, enqueueingThread, forceQueueingSynchronizationWorkaround);
+    this.bufferEnqueuer = new AsynchronousMediaCodecBufferEnqueuer(codec, enqueueingThread);
     this.synchronizeCodecInteractionsWithQueueing = synchronizeCodecInteractionsWithQueueing;
     this.state = STATE_CREATED;
   }